+++ glean/src/index/webext.ts
@@ -7,6 +7,7 @@ import type { ConfigurationInterface } from "../core/config.js";
 
 import platform from "../platform/webext/index.js";
 
+export { ErrorType } from "../core/error/error_type.js";
 export default {
   /**
    * Initialize Glean. This method should only be called once, subsequent calls will be no-op.
@@ -73,15 +74,6 @@ export default {
   },
 
   /**
-   * Unsets the `debugViewTag` debug option.
-   *
-   * This is a no-op is case there is no `debugViewTag` set at the moment.
-   */
-  unsetDebugViewTag(): void {
-    Glean.unsetDebugViewTag();
-  },
-
-  /**
    * Sets the `sourceTags` debug option.
    *
    * Ping tags will show in the destination datasets, after ingestion.
@@ -95,15 +87,6 @@ export default {
    */
   setSourceTags(value: string[]): void {
     Glean.setSourceTags(value);
-  },
-
-  /**
-   * Unsets the `sourceTags` debug option.
-   *
-   * This is a no-op is case there are no `sourceTags` set at the moment.
-   */
-  unsetSourceTags(): void {
-    Glean.unsetSourceTags();
   },
 
   /**