+++ glean/tests/unit/platform/storage.spec.ts
@@ -2,18 +2,51 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import type { WebDriver } from "selenium-webdriver";
 import assert from "assert";
+import sqlite3 from "sqlite3";
 
-import { setupFirefox, webExtensionAPIProxyBuilder } from "./utils/webext";
+import { firefoxDriver, setupFirefox, webExtensionAPIProxyBuilder } from "./utils/webext";
 import type Store from "../../../src/core/storage";
 
 import TestStore from "../../../src/platform/test/storage";
+import QMLStore from "../../../src/platform/qt/storage";
 import WebExtStore from "../../../src/platform/webext/storage";
 import type { JSONValue } from "../../../src/core/utils";
 import { isUndefined } from "../../../src/core/utils";
 
-let firefox: WebDriver;
+/**
+ * QMLStore implementation, but instead of relying on QML's LocalStorage
+ * executes queries using the Node.js sqlite3 package.
+ */
+const TEST_DATABASE_NAME = "TestGlean";
+let QMLMockDB: sqlite3.Database;
+await new Promise<void>(resolve =>
+  QMLMockDB = new sqlite3.Database(`/tmp/${TEST_DATABASE_NAME}.sqlite3`, () => resolve())
+);
+class MockQMLStore extends QMLStore {
+  constructor(tableName: string) {
+    super(tableName, TEST_DATABASE_NAME);
+    this.initialized = Promise.all([super.initialized, this.delete([])]);
+  }
+
+  _executeQuery(query: string): Promise<LocalStorage.QueryResult | undefined> {
+    return new Promise((resolve, reject) => {
+      QMLMockDB.all(query, (err, rows) => {
+        if (err) {
+          console.error("SQLITE ERROR!", err, query);
+          reject();
+        }
+        resolve({
+          rows: {
+            length: rows?.length || 0,
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-return
+            item: (index: number) => rows[index] || undefined
+          }
+        });
+      });
+    });
+  }
+}
 
 // This object will contain the store names and
 // a function that will initialize and return the store when done.
@@ -27,41 +60,46 @@ const stores: {
   "TestStore": {
     initializeStore: (): TestStore => new TestStore("unused")
   },
+  "QMLStore": {
+    initializeStore: (): MockQMLStore => new MockQMLStore("test"),
+    afterAll: async () => {
+      QMLMockDB.close();
+      return Promise.resolve();
+    }
+  },
   "WebExtStore": {
     initializeStore: (): WebExtStore => new WebExtStore("test"),
     before: async () => {
-      if (!firefox) {
-        firefox = await setupFirefox(true);
-        // Browser needs to be global so that WebExtStore will be built and able to use it.
-        global.browser = {
-          storage: {
-            // We need to ignore type checks because TS will complain about
-            // not defining the `remove` method, which is not necessary for our tests.
+      await setupFirefox();
+      // Browser needs to be global so that WebExtStore will be built and able to use it.
+      global.browser = {
+        storage: {
+          // We need to ignore type checks because TS will complain about
+          // not defining the `remove` method, which is not necessary for our tests.
+          //
+          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+          // @ts-ignore
+          local: {
+            // We need to ignore type checks for the following properties because they do not
+            // match perfectly with what is decribed by out web ext types package.
+            // Moreover, it will also complain about not defining the `clear` and `remove`
+            // methods, but these are not necessary for our tests.
             //
             // eslint-disable-next-line @typescript-eslint/ban-ts-comment
             // @ts-ignore
-            local: {
-              // We need to ignore type checks for the following properties because they do not
-              // match perfectly with what is decribed by out web ext types package.
-              // Moreover, it will also complain about not defining the `clear` and `remove`
-              // methods, but these are not necessary for our tests.
-              //
-              // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-              // @ts-ignore
-              get: webExtensionAPIProxyBuilder(firefox, ["storage", "local", "get"]),
-              // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-              // @ts-ignore
-              set: webExtensionAPIProxyBuilder(firefox, ["storage", "local", "set"]),
-              // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-              // @ts-ignore
-              clear: webExtensionAPIProxyBuilder(firefox, ["storage", "local", "clear"])
-            }
+            get: webExtensionAPIProxyBuilder(firefoxDriver, ["storage", "local", "get"]),
+            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+            // @ts-ignore
+            set: webExtensionAPIProxyBuilder(firefoxDriver, ["storage", "local", "set"]),
+            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+            // @ts-ignore
+            clear: webExtensionAPIProxyBuilder(firefoxDriver, ["storage", "local", "clear"])
           }
-        };
-      }
+        }
+      };
       await browser.storage.local.clear();
     },
-    afterAll: async () => await firefox.quit()
+    afterAll: async () => await firefoxDriver.quit()
   }
 };